+++ pages/2_Understand_the_Model.py
@@ -1,6 +1,6 @@
 import streamlit as st
 from sibylapp.view.utils import filtering, setup
-from sibylapp.view import explore_feature, feature_importance, feature_distributions
+from sibylapp.view import explore_feature, feature_importance
 from sibylapp.compute.context import get_term
 
 
@@ -11,10 +11,9 @@ setup.setup_page()
 filtering.view()
 
 
-tab1, tab2, tab3 = st.tabs(
+tab1, tab2 = st.tabs(
     [
         get_term("Feature Importance"),
-        "Summary by Prediction",
         "Explore a %s" % get_term("Feature"),
     ]
 )
@@ -22,7 +21,4 @@ with tab1:
     features = feature_importance.view()
 
 with tab2:
-    feature_distributions.view(features)
-
-with tab3:
     explore_feature.view(features)
